refactor: CrawlingLog 저장 로직 Facade로 분리 - (#314)

+++ detector/src/main/java/com/back/detector/domain/CrawlingLog.java
@@ -22,7 +22,7 @@ public class CrawlingLog extends BaseTimeEntity {
     private String ipAddress;
 
     @Column(nullable = false)
-    private Long requestCount;
+    private int requestCount;
 
     @Column(nullable = false)
     private int timeWindowMinutes;
